requirements.txt
scikit-learn
python-pptx
python-docx
PyPDF2
joblib
tqdm
jieba

